refactor: rename app.py
