+++ tests/authors/test_authors.py
@@ -2,7 +2,7 @@ from unittest import mock
 from fastapi.testclient import TestClient
 import pytest
 
-from app.main import app
+from app.app import app
 from app.modules.authors.repository import AuthorRepository
 from app.modules.authors.routes import get_author_repository
 from ..stubs.author_stubs import authors_entities_stub, authors_schemas_stub
